#580 Institution change of lectureseries does not change video institution
Two other problems/issues came up while working on this issues #580.

1. If termId of a Lectureseries was changed, the termIds of the videos,
which are assigned to the Lectureseries, were not changed either.

2. If the lectureseries of a video is changed, the table
LG_Lectureseries_Creator has not been updated. But the table must be for
the old and new choosen lectureseries, since the creatorId in the table
LG_Lectureseries_Creator dervie from the videos.