Change handling of downloadable/ html5-embedable videos
Until now the uploaded file was used for downloads and
html5-native-player-embeds. Since there is now a smil-file integration
with videos with different bitrates, the smil file will be used to
determine a file with a reasonable rate to be downloaded. 

If no smil file is found, nothing changes, the uploaded file will be
served.
If there is a smil file the file with the bitrate which is nearest to
the target bitrate (predefined as 1,4 mbit) is used.